backend/requirements.txt
fastapi==0.136.1
numpy==2.4.4
onnxruntime==1.26.0
opencv_python==4.13.0.92
Pillow==12.2.0
pycocotools==2.0.11
pydantic==2.13.4
torch==2.11.0
torchvision==0.26.0
ultralytics==8.4.33
